% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCytoscape.R
\name{createCytoGraph}
\alias{createCytoGraph}
\title{Create Graphs of significant interactions}
\usage{
createCytoGraph(enrichment, ents, rels, DEGs, p.thresh=0.05,
                       fc.thresh = log(1.5), numProt=5, ids=NA,
                       numTargets=10)
}
\arguments{
\item{enrichment}{The enrichment tables from the analysis pipeline.  Can be a list
of lists (multiple methods and conditions) or a list (multiple conditions, one
method), or a single data frame}

\item{ents}{The ents table used in the enrichment, must be a single data frame}

\item{rels}{The rels table used in the enrichment, must be a single data frame}

\item{DEGs}{The differentially expressed gene tables, can be a single data frame
or list of data frames matching that which the pipeline was run on.}

\item{p.thresh}{The p value threshold used in determining the significant
differentailly expressed genes for enrichment analysis}

\item{fc.thresh}{The fold count threshold used in determining the significant
differentially expressed genes for enrichment analysis}

\item{numProt}{The number of protiens, ranked by p.value from enrichment, to include}

\item{ids}{Row indices of the regulators you wish to be included in the plot}

\item{numTargets}{The number of targets of each top regulator to show.  Please note that the
number of targets plotted may be smaller than you expect, due to overlap.}
}
\value{
Opens a browser window with the graph, green = proteins, purple = mRNA
}
\description{
Creates graphs of protein-gene interactions based on enrichment analyses
by this pipeline
}
\examples{

ChIP1ap <- filterChIPAtlas(1, NA, "auto", NA, "prostate", NA, NA, FALSE)
files <- list.files("./", ".txt")
degs <- lapply(files, function(x) {
   read.table(x, header=T, sep="\\t") } )
names(degs) <- files
methods <- c("Ternary", "Quaternary", "Enrichment", "Fisher")
enrichment <- runInferenceModels(NULL, NULL, DEGs=degs,
                                method = methods,
                                ents=ChIP1ap$filteredChIP.ents,
                                rels=ChIP1ap$filteredChIP.rels,
                                useFile=F, useMart=TRUE, useBHLH=TRUE,
                                martFN="../CIE/data/mart_human_TFs.csv",
                                BHLHFN="../CIE/data/BHLH_TFs.txt")
createCytoGraph(enrichment, ChIP1ap$filteredChIP.ents,
                ChIP1ap$filteredChIP.rels, degs)
}
