% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterChIP.R
\name{filterChIPAtlas}
\alias{filterChIPAtlas}
\title{filterChIPAtlas}
\usage{
readChIPAtlas(distance, cutoff, cutoffType, cellLines = NA, cellLineType=NA, cellLineDiagnosis = NA, outFileName = NA, writeToFile=TRUE, databaseDir=NA)
}
\arguments{
\item{distance}{A number, either 1, 5, or 10 which indicates what distance (in kb)
from the transcription start site should be considered.}

\item{cutoff}{A number to be used in evaluating which targets have a sufficient binding score.
Binding score filtering depends on the next parameter, cutoffType}

\item{cutoffType}{One of "average", "min", "max", or "auto".  If "average" is selected, only
targets which have an average binding score above the cutoff will be included.  If "min" or
"max" is selected, the binding scores from individual experiments will be evaluated to see
if they are above or below that threshold, respectively.  If none are, the target will not be
included.  If "auto" is selected, the 75th quantile of experiment-specific bindings scores for
each transcription factor will be determined.  Only targets with a bindings score above the
75th quantile will be included.}

\item{cellLines}{A vector of strings which are names of cell lines.  Only interactions which
have been documented in that cell line will be filtered by cutoff and then returned.
It is set to NA by default, and is an optional parameter.}

\item{cellLineType}{A single string which corresponds to a name of an organ.  Only interactions
which have been documented in cell lines originating from that tissue will be filtered by
cutoff and then returned. It is set to NA by default, and is an optional parameter.}

\item{cellLineDiagnosis}{A single string which corresponds to the diagnosis of the person the
cell line is from.  Only interactions which have been documented in cell lines originating from
a person with that diagnosis will be filtered by cutoff and then returned. It is set to NA by
default, and is an optional parameter.}

\item{outFileName}{A string of the desired output file name, optional.  It should be
the part of the file name before the extension (ex. "filename" becomes filename.rels).}

\item{writeToFile}{A boolean value, determining whether the output will be returned to
the environment or written to file (FALSE and TRUE, respectively.  Defaults to TRUE}

\item{databaseDir}{The path to the folder containing cellLines.rds and chip-atlas-*kb.rds.  The
function will search the current working directory if a database directory  is not provided.}

\item{annotate}{Use annotations, defaults to false}

\item{expectProgressObject}{optional parameter specifying if a Shiny progress
bar will be passed}

\item{progressObject}{optional parameter providing the Shiny progress bar}
}
\value{
Two dataframes, differentiated by the suffix to their names both in a returned object
and  when written to file. One, ending in .ents, is the entries from the database which passed
the filters specified.  The other, ending in .rels describes the relations between the entries.
If returned to the environment, the ents table is accessible by [outVarName]$filteredChIP.ents
and the rels table is accessible by [outVarName]$filteredChIP.rels
}
\description{
Given selection parameters, retrieve a portion of the ChIP Atlas database in the
format of an .ents and a .rels file, which are useful for network and graph generation.  This
is also the format required by the runCIE and createCytoGraph functions in this library.
}
\examples{
For 5kb distance from the TSS, an average score of 500, and the cell line SEM:
readChIPAtlas(5, 500, "average", "SEM")
For 5kb distance from the TSS, and automatic cutoff
readChIPAtlas(5, NA, "automatic")
readChIPAtlas(5, NA, "auto")
For 1kb distance from the TSS and select cell line by primary tissue type  blood
and automatic cutoff
readChIPAtlas(1, NA, "auto", cellLineType="blood")
Select by 1kb distance, autocutoff, and only cell lines from blood cells in people
diagnosed with leukemia 
ChIP1autoBldLeuk <- filterChIPAtlas(1, NA, "auto", NA, "blood", "leukemia")
ents <- ChIP1autoBldLeuk$filteredChIP.ents
rels <- ChIP1autoBldLeuk$filteredChIP.rels

}
