% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKBs.R
\name{parseChIP}
\alias{parseChIP}
\title{parseChIP}
\usage{
parseChIP(tsv_dir, processToFilterLater=FALSE)
}
\arguments{
\item{tsv_dir}{The directory containing the data pulled from ChIP
Atlas.}

\item{processToFilterLater}{A switch controlling whether the output
will be .ents and .rels tables (FALSE) or a list of data frames
suitable for later filtering by filterChIPAtlas (TRUE).  Defaults
to FALSE}
}
\value{
A list of two data frames (ents and rels) or a list of
dataframes suitable for later processing.
}
\description{
Process unfiltered ChIP Atlas data, pulled into a
folder with a script like the one that can be found in
../inst/shellScripts/.  Can either return a list of data
frames suitable for filtering with filterChIPAtlas (regardless
of species) or .rels and .ents tables suitable for running
enrichment or other purposes.
}
\examples{
## Both these examples assume you pulled the ChIP Atlas data
## and it can be found at the location "../data/chip-atlas-10kb/"

## Example One
chip-atlas-10kb <- parseChIP("../data/chip-atlas-10kb/",
                             processToFilterLater=TRUE)
saveRDS(chip-atlas-10kb, "../data/chip-atlas-10kb.rds")

entsRels <- filterChIPAtlas(10, 500, "average",
                            databaseDir="../data")
View(entsRels$filteredChIP.ents)
View(entsRels$filteredChIP.rels)

## Example Two

entsRels <- parseChIP("../data/chip-atlas-10kb/")
View(entsRels$ents)
View(entsRels$rels)

}
