% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKBs.R
\name{parseStringKB}
\alias{parseStringKB}
\title{parseStringKB}
\usage{
parseTRED(in.file)
}
\arguments{
\item{ents.file}{A tab separated file containing the entries of STRINGdb}

\item{rels.file}{A tab separated file describing the interactions between the entries of
STRINGdb}

\item{verbose}{A boolean value, indicating whether the output should include verbose messaging.
Defaults to FALSE.}
}
\value{
A list of two dataframes, ents and rels.  The ents dataframe is all the entries from
the selected database. The rels table describes the relations between the entries
}
\description{
Process STRINGdb data into a format usable by this enrichment pipeline. Assumes
that if a protein targets a protein it also targets the gene.
}
\examples{

entsRels <- parseStringKB("string.ents", "string.rels")
ents <- entsRels$ents
rels <- entsRels$rels

}
