% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKBs.R
\name{parseTRED}
\alias{parseTRED}
\title{parseTRED}
\usage{
parseTRED(in.file)
}
\arguments{
\item{in.file}{A csv file containing the TRED database.}
}
\value{
A list of two dataframes, ents and rels.  The ents dataframe is all the entries from
the selected database. The rels table describes the relations between the entries
}
\description{
Process TRED data into a format usable by this enrichment pipeline.
}
\examples{

entsRels <- parseTRED("TRED.csv")
ents <- entsRels$ents
rels <- entsRels$rels

}
