% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTRRUST.R
\name{parseTRRUST}
\alias{parseTRRUST}
\title{parseTRRUST}
\usage{
parseTRED(in.file)
}
\arguments{
\item{in.file}{A tab seperated file containing the TRRUST database.}
}
\value{
A list of two dataframes, ents and rels.  The ents dataframe is all the entries from
the selected database. The rels table describes the relations between the entries
}
\description{
Process TRRUST data into a format usable by this enrichment pipeline.
}
\examples{

entsRels <- parseTRRUST("trrust_rawdata.human.tsv")
ents <- entsRels$ents
rels <- entsRels$rels

}
